+++ packages/oc-template-react-compiler/src/lib/reactOCProviderTemplate.ts
@@ -11,6 +11,7 @@ export default function reactOCProviderTemplate({
   import React from 'react';
   import View from '${removeExtension(viewPath)}';
   import { DataProvider } from 'oc-template-react-compiler/dist/utils/useData'
+  import { createRoot, hydrateRoot } from 'react-dom/client';
 
   function OCProvider(props) {
     const { _staticPath, _baseUrl, _componentName, _componentVersion, ...rest } = props;
@@ -32,18 +33,10 @@ export default function reactOCProviderTemplate({
   }
 
   function renderer(props, element, ssr) {
-    if (window.ReactDOM.createRoot) {
-      if (ssr) {
-        window.ReactDOM.hydrateRoot(element, <OCProvider {...props} />);
-      } else {
-        window.ReactDOM.createRoot(element).render(<OCProvider {...props} />);
-      }
+    if (ssr) {
+      hydrateRoot(element, <OCProvider {...props} />);
     } else {
-      if (ssr) {
-        window.ReactDOM.hydrate(<OCProvider {...props} />, element);
-      } else {
-        window.ReactDOM.render(<OCProvider {...props} />, element);
-      }
+      createRoot(element).render(<OCProvider {...props} />);
     }
   }